+++ gacetaelectronica/eslint.config.mjs
@@ -1,21 +1,18 @@
-import { dirname } from "path";
-import { fileURLToPath } from "url";
-import { FlatCompat } from "@eslint/eslintrc";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
+import { FlatCompat } from '@eslint/eslintrc'
+ 
 const compat = new FlatCompat({
-  baseDirectory: __dirname,
-});
-
+  // import.meta.dirname is available after Node.js v20.11.0
+  baseDirectory: import.meta.dirname,
+})
+ 
 const eslintConfig = [
-  ...compat.extends("next/core-web-vitals", "next/typescript"),
-  {
+  ...compat.config({
+    extends: ['next'],
     rules: {
-      "@typescript-eslint/no-explicit-any": "off",  // Desactivar la regla no-explicit-any
+      'react/no-unescaped-entities': 'off',
+      '@next/next/no-page-custom-font': 'off',
     },
-  },
-];
-
-export default eslintConfig;
+  }),
+]
+ 
+export default eslintConfig
