+++ gacetaelectronica/src/app/api/drive/auth/route.ts
@@ -1,11 +1,10 @@
-import { NextRequest, NextResponse } from 'next/server'
+import { NextResponse } from 'next/server'
 
 // Configuración base para Google OAuth (sin googleapis por ahora)
 const GOOGLE_CLIENT_ID = process.env.GOOGLE_CLIENT_ID
-const GOOGLE_CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET
 const REDIRECT_URI = process.env.GOOGLE_REDIRECT_URI || 'http://localhost:3000/api/drive/callback'
 
-export async function GET(req: NextRequest) {
+export async function GET() {
   try {
     if (!GOOGLE_CLIENT_ID) {
       return NextResponse.json(
