+++ gacetaelectronica/src/app/api/drive/token/route.ts
@@ -44,7 +44,7 @@ export async function POST(req: NextRequest) {
   }
 }
 
-export async function GET(req: NextRequest) {
+export async function GET() {
   try {
     // Generar URL de autorización
     const scopes = [
