+++ gacetaelectronica/src/app/private/layout.tsx
@@ -1,6 +1,6 @@
 // src/app/private/layout.tsx
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "../api/auth/[...nextauth]/options"; // Actualizado para importar desde options
 import { redirect } from "next/navigation";
 
 export default async function PrivateLayout({ children }: { children: React.ReactNode }) {
