+++ gacetaelectronica/src/app/publica/guia-articulo/components/submissionCard.tsx
@@ -1,4 +1,3 @@
-import { Send } from "lucide-react";
 
 export default function SubmissionCard() {
   return (
