+++ gacetaelectronica/src/components/redactor/TestPreview.tsx
@@ -4,7 +4,7 @@ import { useState } from "react"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
-import { Image, Video, FileText } from "lucide-react"
+
 
 export default function TestPreview() {
   const [url, setUrl] = useState("")
